Add singularity finding.
Singularities are found by looking for zeros of the f numerator
without the spuaring. The zeros are found with scipy.optimize.brentq.
scipy.optimize.brentq needs to be given an interval in which f reverses
sign. To find these intervals the domain is divided by a specified number
of points and those intervals are examined for sign changes. If f
switches sign more rapidly then the points then zeros will go undetected.