Fix symbolic nu profile derivation, Fix mu_0 error
To changes are made to the symbolic derivation. First nu is substituted
for the actual number at the beginning. There was an error that
developed if nu was left generic.
Second sp.cancel is used to cancel out rs in the safty_factor formula so
that subsituting r with 0 does not result in divide by zero.

mu_0 was divided by instead of multiplied by in several locations in the
nu2 profile.