Remove complex check, fix error in process_dr
Complexity check probably would never trip since right now
singularity_forbenius converts complex results to reals.
If process_dr gets an interval with interval[0] > interval[1] then it
returns an empty intervals_dr and changes the interval[0] to be the same
as interval[1]. This should force the integrator to only evaluate the
frobenius expansion at the endpoint of the interval and skip the
integration.