Improve offset handeling, Add missing_param saving
Missing params added to analyse code better.
To make sure the intervals do not start or end at a singularity compare
the boundaries with np.allclose() to the sing array.
Use summing across boolean comparison ndarrays to check for interval ends
that are close to singularities.
Have to be careful that comparison to empty lists returns True.