Fix calculation of q profile and def of j_core
For the q profile q0 is now explicitly calculated and the user input q0 is
not used. The user q0 is still used to determine j_core and j_skin. Before
if the equations for j_core and j_skin were not correct the q profile
would be wrong at r=0 and ringing would develop due to the discontinuity.

I found a simplier equation relating q0 and j_core. Definition of j_core
is adjusted.