Add d_w map. Use -1 for d_W is None in binary maps
d_W map is needed to see the magnetitude of d_W. If magneitude is close to
zero they may be within integration error or marginally stable.
Set undetermined d_W due to integration errors to -1 in stability arrays
so the points can be indentified as undetermined in the contour plots.