Add modified scipy splines for performance improvement.
Since the code evaluates a lot of splines at the same r, it is faster to prepare the r with np.asarray, np.at_least1d outside the spline evaluations to avoid costly repetition. The origional scipy code was modified by removing the np.asarray and at_least1d statements.